requirements.txt
# ============================================================
# Semantic Knowledge Pipeline - Requirements
# ============================================================
# Phase 1: Ingestion Engine (Scraping + Database)
# ============================================================

# --- Scraping ---
# Playwright: Browser automation library for scraping JavaScript-rendered pages.
# We use the ASYNC version so we can scrape multiple pages concurrently.
playwright==1.52.0

# --- Data Validation ---
# Pydantic: Enforces strict data schemas (types, constraints) at runtime.
# Every piece of data flowing through our pipeline MUST pass through a Pydantic model.
pydantic==2.11.3

# --- Database ---
# SQLAlchemy: Python SQL toolkit and ORM. Lets us define database tables as Python
# classes and query them using Python methods instead of raw SQL strings.
SQLAlchemy==2.0.41

# psycopg2-binary: The PostgreSQL adapter for Python. "Binary" means it comes
# pre-compiled, so you don't need to build from source. SQLAlchemy uses this
# under the hood to talk to PostgreSQL.
psycopg2-binary==2.9.10

# asyncpg: An async PostgreSQL driver. We'll use this later for high-performance
# async database operations alongside Playwright's async scraping.
asyncpg==0.30.0

# --- AI / LLM Orchestration ---
# LangChain: Framework for building LLM-powered applications.
# Provides a universal interface that works with ANY LLM (OpenAI, Ollama, etc.)
# We use it to: define prompts, call the LLM, parse structured output.
langchain>=0.3.0
langchain-core>=0.3.0

# langchain-ollama: LangChain's Ollama integration.
# This is the ADAPTER that lets LangChain talk to our local Gemma 4 model
# running in Ollama. Without this, LangChain wouldn't know how to call Ollama.
langchain-ollama>=0.3.0

# --- REST API ---
# FastAPI: Modern, high-performance web framework for building APIs.
# Auto-generates Swagger docs, validates request/response with Pydantic.
fastapi>=0.135.0

# Uvicorn: ASGI server that runs FastAPI. Handles HTTP requests and passes
# them to our FastAPI app for processing.
uvicorn>=0.34.0

# --- Utilities ---
# python-dotenv: Loads environment variables from a .env file (DB credentials, API keys).
# Keeps secrets out of our source code.
python-dotenv==1.1.0
